requirements-test.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --resolver=backtracking --extra test -c requirements.txt -o requirements-test.txt
annotated-types==0.7.0
    # via
    #   -c requirements.txt
    #   pydantic
attrs==25.3.0
    # via
    #   -c requirements.txt
    #   frictionless
    #   jsonlines
    #   jsonschema
    #   referencing
certifi==2025.4.26
    # via
    #   -c requirements.txt
    #   requests
chardet==5.2.0
    # via
    #   -c requirements.txt
    #   frictionless
charset-normalizer==3.4.2
    # via
    #   -c requirements.txt
    #   requests
ckanapi==4.8
    # via
    #   -c requirements.txt
    #   hdx-python-api
click==8.2.1
    # via
    #   -c requirements.txt
    #   typer
coverage==7.8.2
    # via pytest-cov
defopt==6.4.0
    # via
    #   -c requirements.txt
    #   hdx-python-api
dnspython==2.7.0
    # via
    #   -c requirements.txt
    #   email-validator
docopt==0.6.2
    # via
    #   -c requirements.txt
    #   ckanapi
    #   num2words
docutils==0.21.2
    # via
    #   -c requirements.txt
    #   defopt
email-validator==2.2.0
    # via
    #   -c requirements.txt
    #   hdx-python-api
et-xmlfile==2.0.0
    # via
    #   -c requirements.txt
    #   openpyxl
frictionless==5.18.1
    # via
    #   -c requirements.txt
    #   hdx-python-utilities
hdx-python-api==6.4.4
    # via
    #   -c requirements.txt
    #   hdx-scraper-iati (pyproject.toml)
hdx-python-country==3.9.4
    # via
    #   -c requirements.txt
    #   hdx-python-api
hdx-python-utilities==3.8.7
    # via
    #   -c requirements.txt
    #   hdx-scraper-iati (pyproject.toml)
    #   hdx-python-api
    #   hdx-python-country
humanize==4.12.3
    # via
    #   -c requirements.txt
    #   frictionless
idna==3.10
    # via
    #   -c requirements.txt
    #   email-validator
    #   requests
ijson==3.4.0
    # via
    #   -c requirements.txt
    #   hdx-python-utilities
inflect==7.5.0
    # via
    #   -c requirements.txt
    #   quantulum3
iniconfig==2.1.0
    # via pytest
isodate==0.7.2
    # via
    #   -c requirements.txt
    #   frictionless
jinja2==3.1.6
    # via
    #   -c requirements.txt
    #   frictionless
jsonlines==4.0.0
    # via
    #   -c requirements.txt
    #   hdx-python-utilities
jsonpath-ng==1.7.0
    # via
    #   -c requirements.txt
    #   libhxl
jsonschema==4.24.0
    # via
    #   -c requirements.txt
    #   frictionless
    #   tableschema-to-template
jsonschema-specifications==2025.4.1
    # via
    #   -c requirements.txt
    #   jsonschema
libhxl==5.2.2
    # via
    #   -c requirements.txt
    #   hdx-python-api
    #   hdx-python-country
loguru==0.7.3
    # via
    #   -c requirements.txt
    #   hdx-python-utilities
makefun==1.16.0
    # via
    #   -c requirements.txt
    #   hdx-python-api
markdown-it-py==3.0.0
    # via
    #   -c requirements.txt
    #   rich
marko==2.1.3
    # via
    #   -c requirements.txt
    #   frictionless
markupsafe==3.0.2
    # via
    #   -c requirements.txt
    #   jinja2
mdurl==0.1.2
    # via
    #   -c requirements.txt
    #   markdown-it-py
more-itertools==10.7.0
    # via
    #   -c requirements.txt
    #   inflect
num2words==0.5.14
    # via
    #   -c requirements.txt
    #   quantulum3
openpyxl==3.1.5
    # via
    #   -c requirements.txt
    #   hdx-python-utilities
packaging==25.0
    # via pytest
petl==1.7.16
    # via
    #   -c requirements.txt
    #   frictionless
pluggy==1.6.0
    # via pytest
ply==3.11
    # via
    #   -c requirements.txt
    #   jsonpath-ng
    #   libhxl
pockets==0.9.1
    # via
    #   -c requirements.txt
    #   sphinxcontrib-napoleon
pydantic==2.11.5
    # via
    #   -c requirements.txt
    #   frictionless
pydantic-core==2.33.2
    # via
    #   -c requirements.txt
    #   pydantic
pygments==2.19.1
    # via
    #   -c requirements.txt
    #   pytest
    #   rich
pyphonetics==0.5.3
    # via
    #   -c requirements.txt
    #   hdx-python-utilities
pytest==8.4.0
    # via
    #   hdx-scraper-iati (pyproject.toml)
    #   pytest-cov
pytest-cov==6.1.1
    # via hdx-scraper-iati (pyproject.toml)
python-dateutil==2.9.0.post0
    # via
    #   -c requirements.txt
    #   frictionless
    #   hdx-python-utilities
    #   libhxl
python-io-wrapper==0.3.1
    # via
    #   -c requirements.txt
    #   libhxl
python-slugify==8.0.4
    # via
    #   -c requirements.txt
    #   ckanapi
    #   frictionless
pyyaml==6.0.2
    # via
    #   -c requirements.txt
    #   frictionless
    #   tableschema-to-template
quantulum3==0.9.2
    # via
    #   -c requirements.txt
    #   hdx-python-api
ratelimit==2.2.1
    # via
    #   -c requirements.txt
    #   hdx-python-utilities
referencing==0.36.2
    # via
    #   -c requirements.txt
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   -c requirements.txt
    #   ckanapi
    #   frictionless
    #   hdx-python-api
    #   libhxl
    #   requests-file
requests-file==2.1.0
    # via
    #   -c requirements.txt
    #   hdx-python-utilities
rfc3986==2.0.0
    # via
    #   -c requirements.txt
    #   frictionless
rich==14.0.0
    # via
    #   -c requirements.txt
    #   typer
rpds-py==0.25.1
    # via
    #   -c requirements.txt
    #   jsonschema
    #   referencing
ruamel-yaml==0.18.14
    # via
    #   -c requirements.txt
    #   hdx-python-utilities
ruamel-yaml-clib==0.2.12
    # via
    #   -c requirements.txt
    #   ruamel-yaml
setuptools==80.9.0
    # via
    #   -c requirements.txt
    #   ckanapi
shellingham==1.5.4
    # via
    #   -c requirements.txt
    #   typer
simpleeval==1.0.3
    # via
    #   -c requirements.txt
    #   frictionless
simplejson==3.20.1
    # via
    #   -c requirements.txt
    #   ckanapi
six==1.17.0
    # via
    #   -c requirements.txt
    #   ckanapi
    #   pockets
    #   python-dateutil
    #   sphinxcontrib-napoleon
sphinxcontrib-napoleon==0.7
    # via
    #   -c requirements.txt
    #   defopt
structlog==25.4.0
    # via
    #   -c requirements.txt
    #   libhxl
tableschema-to-template==0.0.13
    # via
    #   -c requirements.txt
    #   hdx-python-utilities
tabulate==0.9.0
    # via
    #   -c requirements.txt
    #   frictionless
tenacity==9.1.2
    # via
    #   -c requirements.txt
    #   hdx-python-country
text-unidecode==1.3
    # via
    #   -c requirements.txt
    #   python-slugify
typeguard==4.4.3
    # via
    #   -c requirements.txt
    #   inflect
typer==0.16.0
    # via
    #   -c requirements.txt
    #   frictionless
typing-extensions==4.14.0
    # via
    #   -c requirements.txt
    #   frictionless
    #   pydantic
    #   pydantic-core
    #   referencing
    #   typeguard
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   -c requirements.txt
    #   pydantic
unidecode==1.4.0
    # via
    #   -c requirements.txt
    #   libhxl
    #   pyphonetics
urllib3==2.4.0
    # via
    #   -c requirements.txt
    #   libhxl
    #   requests
validators==0.35.0
    # via
    #   -c requirements.txt
    #   frictionless
wheel==0.45.1
    # via
    #   -c requirements.txt
    #   libhxl
xlrd==2.0.1
    # via
    #   -c requirements.txt
    #   hdx-python-utilities
xlrd3==1.1.0
    # via
    #   -c requirements.txt
    #   libhxl
xlsx2csv==0.8.4
    # via
    #   -c requirements.txt
    #   hdx-python-utilities
xlsxwriter==3.2.3
    # via
    #   -c requirements.txt
    #   tableschema-to-template
xlwt==1.3.0
    # via
    #   -c requirements.txt
    #   hdx-python-utilities
